hb_web/requirements.txt
FLask>=3.0.0
psycopg2-binary>=2.9.1
bcrypt>=3.2.0
sqlalchemy>=2.0.0
qrcode>=7.0.0